exit fixed

+++ includes/builtins.h
@@ -6,7 +6,7 @@
 /*   By: user <user@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/23 16:59:28 by user              #+#    #+#             */
-/*   Updated: 2024/01/17 15:28:14 by user             ###   ########.fr       */
+/*   Updated: 2024/01/17 16:31:12 by user             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,7 +43,7 @@ void	ft_unset(char **command);
  * @brief	Prints exit message, free what is needed and exit program.
  * @param command a command formed from an array of tokens.
  */
-void	ft_exit(char **command, int status);
+void	ft_exit(char **command, long status);
 
 /**
  * @brief	Prints whats was the input.
